Update the glucose chart view to use different time ranges
This commit includes changes to allow the user to select different
time ranges to display glucose values for the main glucose chart. It
also provides scrolling to see backwards to 24 hours worth of data.

In the iPhone app, there are buttons that people can use to switch. In
the Watch app, they click on the chart to change the time range.

It also includes some minor refactoring to improve the readability of
the code.