Add a "Diagnostics" view
This commit adds a new view to provide some details of the inner
workings of the core algorithm. It includes details on dosing
commands, glucose and IoB, the PID controller, and how much insulin
the ML is adding.

It uses a standard View / View Model architecture drawing from the
ClosedLoopService as the source of truth for algorithm runs.