Loosen restriction on micro bolus
This commit removes the constraint that disables micro bolus when
glucose is going down. Since we want to use a micro bolus to deliver
insulin faster we should use it any time the algorithm decides to
dose even if glucose is falling.

This change will help lower glucose faster overnight.